requirements.txt
streamlit>=1.32.0
python-dotenv>=1.0.0
pymupdf>=1.23.0
python-docx>=0.8.11
langchain-groq>=0.1.0
langchain-community>=0.0.29
sentence-transformers>=2.7.0
faiss-cpu>=1.7.4